added order resource.Order simple test

+++ LykkePay/Models/PostPurchaseModel.cs
@@ -35,16 +35,16 @@ namespace LykkePay.Models
     {
         public double percent { get; set; }
         public int pips { get; set; }
-        public float fixedFee { get; set; }
+        public double fixedFee { get; set; }
 
-        public PostMarkup(double percent, int pips, float fixedFee)
+        public PostMarkup(double percent, int pips, double fixedFee)
         {
             this.percent = percent;
             this.pips = pips;
             this.fixedFee = fixedFee;
         }
 
-        public PostMarkup(MarkupModel markupModel, float fixedFee)
+        public PostMarkup(MarkupModel markupModel, double fixedFee)
         {
             this.percent = markupModel.markup.percent;
             this.pips = markupModel.markup.pips;
